refactor(web): extract organization deletion dialog
